+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ### Changed
-
+- Values - remove duplicated parse-ago
 - Apps - Saving an app should move user to it #991
 - Deployment set - Experimental feature (hidden for prod env)
 - Time - delta minutes bugfix